Create Controlling RPM of Fan
Some code that hijacks Timer2 to generate PWM signals on Pin 3 and or Pin 11 simultaneously with different duty cycle. 

You can comment out the parts for the pin that you don’t need. Put this stuff in setup() and then to set the duty cycle just load it into OCR2A(Pin11) or OCR2B(Pin3) depending.
You load a value from 0 to 255 just like with analogWrite(). 
The pulse rate is kinda high, but let's see if it works before getting more complicated with it. Each duty cycle increment adds 125nS to the pulse width.